backend/requirements.txt
# Backend Dependencies — RSA MVP Enhanced
# ==========================================

# --- Web Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# --- Database ---
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# --- AI / NLP ---
langchain>=0.1.0
langchain-community>=0.0.6
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
openai>=1.6.0
tiktoken>=0.5.0

# --- File Parsing ---
PyPDF2>=3.0.1
python-docx>=1.1.0

# --- Async Tasks ---
celery[redis]>=5.3.0
redis>=5.0.0

# --- Utilities ---
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# --- Security ---
cryptography>=41.0.0

# --- Report Generation ---
reportlab>=4.0.0
